Add native packaging targets and runtime handling logic
The PackNative target takes all the available slang release binaries downloaded with DownloadNativeSlangBinaries and creates runtime nuget packages for each platform. As well as a metapackage that tries to select the correct runtime on restore.
The InitLocalFeed target takes those packages and creates a local nuget feed for development. Only available for local builds.